+++ src/opt/mod.rs
@@ -1,13 +1,13 @@
 mod generate;
 mod greedy;
-mod quad;
 mod rampify;
 mod terrain;
+mod tile_grid;
 mod wedge;
 
 pub use generate::*;
 pub use greedy::*;
-pub use quad::*;
 pub use rampify::*;
 pub use terrain::*;
+pub use tile_grid::*;
 pub use wedge::*;